src/backend/requirements.txt
bcrypt==4.2.1
blinker==1.9.0
boto3
botocore
click==8.1.7
Flask==3.1.0
Flask-JWT-Extended==4.7.1
itsdangerous==2.2.0
Jinja2==3.1.4
MarkupSafe==3.0.2
psycopg==3.2.3
psycopg-c==3.2.3
psycopg2==2.9.10
PyJWT==2.10.1
python-dotenv==1.0.1
typing_extensions==4.12.2
Werkzeug==3.1.3
